configure and systemd

+++ install.sh
@@ -5,34 +5,91 @@ set -xeu
 # https://dist.ipfs.io/
 IPFS_VER=v0.12.2
 
-sudo apt clean all
-sudo apt update
-sudo apt autoremove
-sudo apt update
-sudo apt upgrade
-sudo apt autoremove
+DO_UPDATE=0
 
-sudo apt-get install tar wget jq
 
-cd /tmp
 
-if [[ -d "ipfs-gateway" ]]; then
+if [[ "$DO_UPDATE" == "1" ]]; then
+	sudo apt clean all
+	sudo apt update
+	sudo apt autoremove
+	sudo apt update
+	sudo apt upgrade
+	sudo apt autoremove
+
+	sudo apt-get install tar wget jq
+fi
+
+
+
+if [[ -f "/usr/local/bin/ipfs" ]]; then
+	echo ipfs already exists
+	echo delete /usr/local/bin/ipfs to update
+else
+	cd /tmp
+
+	if [[ -d "ipfs-gateway" ]]; then
+		rm -rfv ipfs-gateway
+	fi
+
+
+	mkdir -p ipfs-gateway
+
+	cd ipfs-gateway
+
+	wget --no-clobber https://dist.ipfs.io/go-ipfs/${IPFS_VER}/go-ipfs_${IPFS_VER}_linux-amd64.tar.gz
+
+	tar xfv go-ipfs_${IPFS_VER}_linux-amd64.tar.gz
+
+	# Move it into your bin. This requires root permissions.
+	sudo cp go-ipfs/ipfs /usr/local/bin/
+
+	cd /tmp
+
 	rm -rfv ipfs-gateway
 fi
 
-mkdir -p ipfs-gateway
 
-cd ipfs-gateway
 
-wget --no-clobber https://dist.ipfs.io/go-ipfs/${IPFS_VER}/go-ipfs_${IPFS_VER}_linux-amd64.tar.gz
+if [[ -d "/home/ipfs" ]]; then
+	echo "ipfs already exists"
+else
+	echo "creating ipfs user"
+	sudo adduser ipfs
+fi
 
-tar xfv go-ipfs_${IPFS_VER}_linux-amd64.tar.gz
 
-# Move it into your bin. This requires root permissions.
-sudo cp go-ipfs/ipfs /usr/local/bin/
+if [[ -f "/etc/systemd/system/ipfs.service" ]]; then
+	echo "systemd already configured: /etc/systemd/system/ipfs.service"
+	echo "RUN: sudo systemctl start ipfs"
+else
+	echo "configuring systemd"
+	cat <<CAT > /tmp/ipfs.service
+[Unit]
+Description=IPFS Daemon
+After=syslog.target network.target remote-fs.target nss-lookup.target
 
-cd /tmp
+[Service]
+Type=simple
+ExecStart=/usr/local/bin/ipfs daemon --enable-namesys-pubsub
+ExecStop=/usr/local/bin/ipfs shutdown
+User=ipfs
+Restart=always
 
-rm -rfv ipfs-gateway
+[Install]
+WantedBy=multi-user.target
+CAT
 
+sudo mv /tmp/ipfs.service /etc/systemd/system/ipfs.service
+
+cat /etc/systemd/system/ipfs.service
+
+sudo systemctl daemon-reload
+sudo systemctl enable ipfs
+#sudo systemctl start ipfs
+sudo systemctl status ipfs
+fi
+
+
+ipfs version
